refactor: move cache to store
