+++ aftermath/internal/store/store.go
@@ -1,4 +1,4 @@
-package cache
+package store
 
 import (
 	"aftermath/internal/bibliography"
